+++ crates/bcr-ebill-web/src/main.rs
@@ -81,6 +81,7 @@ async fn start(
     let keys = db.identity_store.get_key_pair().await?;
     info!("Local node id: {local_node_id:?}");
     info!("Local npub: {:?}", keys.get_nostr_npub()?);
+    info!("Local npriv: {:?}", keys.get_nostr_npriv()?);
     info!("Local npub as hex: {:?}", keys.get_nostr_npub_as_hex());
 
     if db.identity_store.exists().await {
